Refactor FileStorageService to address the type cast related maintainability issue

+++ src/main/java/com/bytebreeze/quickdrop/service/FileStorageService.java
@@ -21,7 +21,7 @@ public class FileStorageService {
 	@Value("${storage.local.path}")
 	private String localUploadPath;
 
-	private static final long MAX_FILE_SIZE = 5 * 1024 * 1024;
+	private static final long MAX_FILE_SIZE = (long) 5 * 1024 * 1024;
 
 	// Allowed file extensions
 	private static final List<String> ALLOWED_EXTENSIONS = Arrays.asList("jpg", "jpeg", "png", "pdf");
